refactor: reorganize API structure and update sensor routes to v1
